+++ src/featureslist.h
@@ -106,13 +106,13 @@ FEATURE_SIMPLE(inAppAuthentication,      // Feature ID
                QStringList(),            // feature dependencies
                FeatureCallback_true)
 
-FEATURE_SIMPLE(inAppProductSelection,       // Feature ID
-               "In-app Product Selection",  // Feature name
-               "2.12",                      // released
-               FeatureCallback_true,        // Can be flipped on
-               FeatureCallback_true,        // Can be flipped off
-               QStringList(),               // feature dependencies
-               FeatureCallback_inAppProductSelection)
+FEATURE_SIMPLE(inAppPurchase,         // Feature ID
+               "In-app Purchase",     // Feature name
+               "2.4",                 // released
+               FeatureCallback_true,  // Can be flipped on
+               FeatureCallback_true,  // Can be flipped off
+               QStringList(),         // feature dependencies
+               FeatureCallback_inAppPurchase)
 
 FEATURE_SIMPLE(keyRegeneration,       // Feature ID
                "Key Regeneration",    // Feature name
