+++ src/purchasewebhandler.h
@@ -2,8 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#ifndef PurchaseWebHandler_H
-#define PurchaseWebHandler_H
+#ifndef PURCHASEWEBHANDLER_H
+#define PURCHASEWEBHANDLER_H
 
 #include "purchasehandler.h"
 
@@ -15,11 +15,14 @@ class PurchaseWebHandler final : public PurchaseHandler {
   explicit PurchaseWebHandler(QObject* parent);
   ~PurchaseWebHandler();
 
+  Q_INVOKABLE void subscribe(
+      const QString& productIdentifier = QString()) override;
+
+ public slots:
+  void stopSubscription();
+
  protected:
-  void nativeRegisterProducts() override;
-  void nativeStartSubscription(Product* product) override;
-  void nativeRestoreSubscription() override;
-  void startSubscription(const QString& productIdentifier);
+  void startSubscription();
 };
 
-#endif  // PurchaseWebHandler_H
+#endif  // PURCHASEWEBHANDLER_H
